# Conflicts:
#	Content.Client/Medical/CrewMonitoring/CrewMonitoringWindow.xaml.cs
#	Content.Client/Shuttles/UI/DockingControl.cs
#	Content.Client/Shuttles/UI/RadarControl.cs
#	Content.Server/Medical/SuitSensors/SuitSensorComponent.cs
#	Content.Server/Medical/SuitSensors/SuitSensorSystem.cs
#	Content.Shared/Medical/SuitSensor/SharedSuitSensor.cs

+++ Content.Server/Explosion/EntitySystems/ExplosionSystem.TileFill.cs
@@ -153,7 +153,7 @@ public sealed partial class ExplosionSystem : EntitySystem
                 if (tilesInIteration[i] * intensityIncrease >= remainingIntensity)
                 {
                     // there is not enough intensity left to distribute. add a fractional amount and break.
-                    iterationIntensity[i] += (float) remainingIntensity / tilesInIteration[i];
+                    iterationIntensity[i] += remainingIntensity / tilesInIteration[i];
                     remainingIntensity = 0;
                     break;
                 }
@@ -219,7 +219,7 @@ public sealed partial class ExplosionSystem : EntitySystem
             tilesInIteration.Add(newTileCount);
             if (newTileCount * stepSize >= remainingIntensity)
             {
-                iterationIntensity.Add((float) remainingIntensity / newTileCount);
+                iterationIntensity.Add(remainingIntensity / newTileCount);
                 break;
             }
